backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.28.1
openai==1.58.1
anthropic==0.42.0
google-generativeai==0.8.4
apscheduler==3.10.4
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
aiosqlite==0.20.0
